Release 4.1.0 - Nuget. Sitemap
it is now possible to automatically generate and deliver sitemap.xml directly via a link. How it works: all paths requested by clients are stored in memory. When the route for retrieving the XML sitemap is requested, the system accesses the stored data to obtain the list of URLs, generates the sitemap on the fly, and immediately sends it to the client.